refactor(#2163): relocate false-negative verify to data_access/presence_verify.rs; fmt
